[Feat] core:database 모듈 생성 dao, database, entity 등 이동
