+++ core/database/src/main/java/com/chan/database/AppDatabase.kt
@@ -1,16 +1,16 @@
-package com.chan.home.data
+package com.chan.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.chan.home.data.dao.HomeBannerDao
-import com.chan.home.data.dao.HomePopularItemDao
-import com.chan.home.data.dao.HomeSaleProductDao
-import com.chan.home.data.dao.RankingCategoryDao
-import com.chan.home.data.entity.home.HomeBannerEntity
-import com.chan.home.data.entity.home.HomePopularItemEntity
-import com.chan.home.data.entity.home.HomeSaleProductEntity
-import com.chan.home.data.entity.ranking.RankingCategoryEntity
+import com.chan.database.dao.HomeBannerDao
+import com.chan.database.dao.HomePopularItemDao
+import com.chan.database.dao.HomeSaleProductDao
+import com.chan.database.dao.RankingCategoryDao
+import com.chan.database.entity.home.HomeBannerEntity
+import com.chan.database.entity.home.HomePopularItemEntity
+import com.chan.database.entity.home.HomeSaleProductEntity
+import com.chan.database.entity.ranking.RankingCategoryEntity
 
 @Database(
     entities = [HomeBannerEntity::class, HomePopularItemEntity::class, RankingCategoryEntity::class, HomeSaleProductEntity::class],
