+++ core/database/src/main/java/com/chan/database/Converters.kt
@@ -1,21 +1,24 @@
-package com.chan.home.data
+package com.chan.database
 
 import androidx.room.TypeConverter
+import com.chan.database.entity.ranking.RankingCategoryEntity.RankingCategoryItems
 import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
+import kotlin.text.isNullOrEmpty
 
 class Converters {
     private val gson = Gson()
 
     @TypeConverter
-    fun itemsToJson(items: List<com.chan.home.data.entity.ranking.RankingCategoryEntity.RankingCategoryItems>): String =
+    fun itemsToJson(items: List<RankingCategoryItems>): String =
         gson.toJson(items)
 
     @TypeConverter
-    fun jsonToItems(json: String): List<com.chan.home.data.entity.ranking.RankingCategoryEntity.RankingCategoryItems> =
+    fun jsonToItems(json: String): List<RankingCategoryItems> =
         gson.fromJson(
             json,
             object :
-                com.google.gson.reflect.TypeToken<List<com.chan.home.data.entity.ranking.RankingCategoryEntity.RankingCategoryItems>>() {}.type
+                TypeToken<List<RankingCategoryItems>>() {}.type
         )
 
     @TypeConverter
@@ -26,7 +29,7 @@ class Converters {
     @TypeConverter
     fun toStringList(value: String?): List<String>? {
         if (value.isNullOrEmpty()) return emptyList()
-        val type = object : com.google.gson.reflect.TypeToken<List<String>>() {}.type
+        val type = object : TypeToken<List<String>>() {}.type
         return gson.fromJson(value, type)
     }
 }
