+++ core/database/src/main/java/com/chan/database/dao/HomeBannerDao.kt
@@ -1,10 +1,10 @@
-package com.chan.home.data.dao
+package com.chan.database.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.chan.home.data.entity.home.HomeBannerEntity
+import com.chan.database.entity.home.HomeBannerEntity
 
 @Dao
 interface HomeBannerDao {
