+++ core/database/src/main/java/com/chan/database/entity/home/HomeBannerEntity.kt
@@ -1,0 +1,11 @@
+package com.chan.database.entity.home
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "homeBanner")
+data class HomeBannerEntity(
+    @PrimaryKey val id: Int,
+    val imageUrl: String,
+    val title: String
+)
