+++ feature/home/src/main/java/com/chan/home/data/repository/HomeSaleProductRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.chan.home.data.repository
 
-import com.chan.home.data.dao.HomeSaleProductDao
+import com.chan.database.dao.HomeSaleProductDao
 import com.chan.home.data.datasource.HomeSaleProductSource
-import com.chan.home.data.entity.home.toDomain
+import com.chan.home.data.mapper.toDomain
 import com.chan.home.domain.repository.HomeSaleProductRepository
 import com.chan.home.domain.vo.HomeSaleProductVO
 import javax.inject.Inject
