+++ settings.gradle.kts
@@ -30,3 +30,4 @@ include(":feature:category")
 
 include(":core:android")
 
+include(":core:database")
